Add a fixture that initiates a singleton Database instance
- Also, move clean_db, database fixture in common.
